[fixup][table-runtime] Moved functions to join.interval package
