+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/stream/StreamExecIntervalJoin.java
@@ -43,12 +43,12 @@ import org.apache.flink.table.runtime.generated.GeneratedJoinCondition;
 import org.apache.flink.table.runtime.keyselector.RowDataKeySelector;
 import org.apache.flink.table.runtime.operators.join.KeyedCoProcessOperatorWithWatermarkDelay;
 import org.apache.flink.table.runtime.operators.join.OuterJoinPaddingUtil;
-import org.apache.flink.table.runtime.operators.join.PaddingLeftMapFunction;
-import org.apache.flink.table.runtime.operators.join.PaddingRightMapFunction;
+import org.apache.flink.table.runtime.operators.join.interval.FilterAllFlatMapFunction;
 import org.apache.flink.table.runtime.operators.join.interval.IntervalJoinFunction;
+import org.apache.flink.table.runtime.operators.join.interval.PaddingLeftMapFunction;
+import org.apache.flink.table.runtime.operators.join.interval.PaddingRightMapFunction;
 import org.apache.flink.table.runtime.operators.join.interval.ProcTimeIntervalJoin;
 import org.apache.flink.table.runtime.operators.join.interval.RowTimeIntervalJoin;
-import org.apache.flink.table.runtime.operators.misc.FilterAllFlatMapFunction;
 import org.apache.flink.table.runtime.typeutils.InternalTypeInfo;
 import org.apache.flink.table.types.logical.RowType;
 import org.apache.flink.util.Preconditions;
