+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/operators/join/interval/FilterAllFlatMapFunction.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.runtime.operators.misc;
+package org.apache.flink.table.runtime.operators.join.interval;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.common.functions.FlatMapFunction;
